Fixed GUnicorn error unable to open camera. Changed video "source" from '0' OR 'sys.argv[1]' to only '0' (indicating /dev/video0).
The "source" of the camera, before, was sys.argv[1] if sys.srgv[1]
existed, otherwise it was /dev/video0. This is fine when prototyping
with Flask, since we typically do not include a parameter for
sys.argv[1] (when running 'python __init__.py <could put source
here>'). However, gunicorn DOES have command-line arguments and
sys.argv[1] is NOT a video source. For this reason, it couln't read
frames from sys.argv[1].