Frame & Annotation Selection.
- Fixed inability for gunicorn to handle more than 1 client by
increasing --threads to 10 (as opposed to default 1)

- Added a sidebar for the camera stream that allows the user to select
the frame and annotations to be displayed.

- the frame Subject now starts a thread for itself to capture opencv
frame, and it starts a thread for each of its observers. Each observer
thread uses the subject's current frame. This is meant to allow frames
to be captured and processed seperately.

- Refactored modules running as scripts (frame subject, laneModel).

- A frame Model now exists to keep track of Nodes that have been added
to it. Example usage: laneModel.py